fix: extract RunTimeline pure logic to a .ts module for node-env tests
tests/web/run-timeline.test.ts imported RunTimeline.tsx, which Vitest's node
environment can't parse (no JSX transform), failing the suite. Moved the types
+ currentPhase/sortRuns into components/run-timeline.ts (mirroring the
approvals.ts / runTrace.ts split); the .tsx re-exports them. Full suite green
(213 files, 2160 tests).
